requirements.txt
python-binance==1.0.19
python-dotenv==1.0.1
pandas==2.2.2
watchdog==4.0.0
prometheus_client==0.20.0
plotly==5.23.0